test(ui): split child-session cap cases under the max-lines budget
